Hot reload issue handling
