+++ alias/__init__.py
@@ -1,7 +1,10 @@
+import importlib
 from cog_shared.sinbad_libs import extra_setup
-from .alias import AliasRewrite
+
+from . import alias
 
 
 @extra_setup
 def setup(bot):
-    bot.add_cog(AliasRewrite(bot))
+    module = importlib.reload(alias)
+    bot.add_cog(module.AliasRewrite(bot))
