+++ antimentionspam/__init__.py
@@ -1,7 +1,10 @@
+import importlib
+
 from . import antimentionspam
 from cog_shared.sinbad_libs import extra_setup
 
 
 @extra_setup
 def setup(bot):
-    bot.add_cog(antimentionspam.AntiMentionSpam(bot))
+    module = importlib.reload(antimentionspam)
+    bot.add_cog(module.AntiMentionSpam(bot))
