+++ bansync/__init__.py
@@ -1,7 +1,10 @@
-from . import bansync
+import importlib
+from . import bansync, converters
 from cog_shared.sinbad_libs import extra_setup
 
 
 @extra_setup
 def setup(bot):
-    bot.add_cog(bansync.BanSync(bot))
+    importlib.reload(converters)
+    module = importlib.reload(bansync)
+    bot.add_cog(module.BanSync(bot))
