+++ devtools/__init__.py
@@ -1,7 +1,10 @@
+import importlib
+
 from . import core
 from cog_shared.sinbad_libs import extra_setup
 
 
 @extra_setup
 def setup(bot):
-    bot.add_cog(core.DevTools(bot))
+    module = importlib.reload(core)
+    bot.add_cog(module.DevTools(bot))
