+++ economytrickle/__init__.py
@@ -1,7 +1,12 @@
-from . import core
+import importlib
+
+from . import core, activity, converters
 from cog_shared.sinbad_libs import extra_setup
 
 
 @extra_setup
 def setup(bot):
-    bot.add_cog(core.EconomyTrickle(bot))
+    importlib.reload(activity)
+    importlib.reload(converters)
+    module = importlib.reload(core)
+    bot.add_cog(module.EconomyTrickle(bot))
