+++ embedmaker/__init__.py
@@ -1,7 +1,12 @@
-from . import embedmaker
+import importlib
+
+from . import embedmaker, serialize, yaml_parse
 from cog_shared.sinbad_libs import extra_setup
 
 
 @extra_setup
 def setup(bot):
-    bot.add_cog(embedmaker.EmbedMaker(bot))
+    importlib.reload(serialize)
+    importlib.reload(yaml_parse)
+    module = importlib.reload(embedmaker)
+    bot.add_cog(module.EmbedMaker(bot))
