+++ guildblacklist/__init__.py
@@ -1,7 +1,10 @@
+import importlib
+
 from . import guildblacklist
 from cog_shared.sinbad_libs import extra_setup
 
 
 @extra_setup
 def setup(bot):
-    bot.add_cog(guildblacklist.GuildBlacklist(bot))
+    module = importlib.reload(guildblacklist)
+    bot.add_cog(module.GuildBlacklist(bot))
