+++ guildwhitelist/__init__.py
@@ -1,7 +1,10 @@
+import importlib
+
 from . import guildwhitelist
 from cog_shared.sinbad_libs import extra_setup
 
 
 @extra_setup
 def setup(bot):
-    bot.add_cog(guildwhitelist.GuildWhitelist(bot))
+    module = importlib.reload(guildwhitelist)
+    bot.add_cog(module.GuildWhitelist(bot))
