+++ quotetools/__init__.py
@@ -1,7 +1,11 @@
-from . import quotetools
+import importlib
+
+from . import quotetools, helpers
 from cog_shared.sinbad_libs import extra_setup
 
 
 @extra_setup
 def setup(bot):
-    bot.add_cog(quotetools.QuoteTools(bot))
+    importlib.reload(helpers)
+    module = importlib.reload(quotetools)
+    bot.add_cog(module.QuoteTools(bot))
